pyproject.toml
[tool.poetry]
name = "mednotice"
version = "0.1.0"
description = ""
authors = ["Praze-hub <godspraiseokechukwu07@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
django = ">=4.2,<5.0"
django-environ = ">=0.12.0,<0.13.0"
psycopg = {extras = ["binary"], version = "^3.3.2"}
celery = "^5.6.2"
redis = "^7.1.0"
djangorestframework = "^3.16.1"
drf-spectacular = "^0.29.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
